Fix signature verification bug
I fixed the bug in the Message get_signature_string method which
was causing tests to fail. The hexlify function requires a bytes
object, and it was being passed self.message, which is of type
str. I just added .encode() to the end to convert it to bytes.

I also reformatted the docstring of the keys.py module.