Fix mining interrupts
Previously, if we had more than one mining thread, receiving a new block would
only interrupt one of them, and only after the block was added. I changed this
so all mining threads are interrupted when we start adding the block and only
resume after we finish adding the new block.